+++ fusionpbx/app/gateways/gateway_delete.php
@@ -52,38 +52,31 @@ if (strlen($_GET["id"])>0) {
 		$prep_statement->execute();
 		$result = $prep_statement->fetchAll(PDO::FETCH_NAMED);
 		foreach ($result as &$row) {
+			$gateway_uuid = $row["gateway_uuid"];
 			$gateway = $row["gateway"];
 			$profile = $row["profile"];
-			break; //limit to 1 row
 		}
 		unset ($prep_statement);
 
 	//delete the xml file
-		if (count($_SESSION["domains"]) > 1) {
-			$gateway_xml_file = $_SESSION['switch']['gateways']['dir']."/".$profile."/v_".$_SESSION['domain_name'].'-'.$gateway.".xml";
+		if ($_SESSION['switch']['sip_profiles']['dir'] != '') {
+			$gateway_xml_file = $_SESSION['switch']['sip_profiles']['dir']."/".$profile."/v_".$gateway_uuid.".xml";
+			if (file_exists($gateway_xml_file)) {
+				unlink($gateway_xml_file);
+			}
 		}
-		else {
-			$gateway_xml_file = $_SESSION['switch']['gateways']['dir']."/".$profile."/v_".$gateway.".xml";
-		}
-		unlink($gateway_xml_file);
 
 	//create the event socket connection and stop the gateway
 		if (!$fp) {
 			$fp = event_socket_create($_SESSION['event_socket_ip_address'], $_SESSION['event_socket_port'], $_SESSION['event_socket_password']);
 		}
-		if ($fp) {
-			//send the api gateway stop command over event socket
-				if (count($_SESSION["domains"]) > 1) {
-					$tmp_cmd = 'api sofia profile '.$profile.' killgw '.$_SESSION['domain_name'].'-'.$gateway;
-				}
-				else {
-					$tmp_cmd = 'api sofia profile '.$profile.' killgw '.$gateway;
-				}
-				$response = event_socket_request($fp, $tmp_cmd);
-				unset($tmp_cmd);
-		}
 
-	//delete gateway
+	//send the api gateway stop command over event socket
+		$cmd = 'api sofia profile '.$profile.' killgw '.$gateway_uuid;
+		$response = event_socket_request($fp, $cmd);
+		unset($cmd);
+
+	//delete the gateway
 		$sql = "delete from v_gateways ";
 		$sql .= "where domain_uuid = '$domain_uuid' ";
 		$sql .= "and gateway_uuid = '$id' ";
@@ -93,13 +86,11 @@ if (strlen($_GET["id"])>0) {
 	//syncrhonize configuration
 		save_gateway_xml();
 
-	//synchronize the xml config
-		save_dialplan_xml();
-
 	//delete the gateways from memcache
 		$fp = event_socket_create($_SESSION['event_socket_ip_address'], $_SESSION['event_socket_port'], $_SESSION['event_socket_password']);
 		if ($fp) {
-			$switch_cmd = "memcache delete configuration:sofia.conf";
+			$hostname = trim(event_socket_request($fp, 'api switchname'));
+			$switch_cmd = "memcache delete configuration:sofia.conf:".$hostname;
 			$switch_result = event_socket_request($fp, 'api '.$switch_cmd);
 		}
 
@@ -110,9 +101,9 @@ if (strlen($_GET["id"])>0) {
 			}
 			if ($fp) {
 				//send the api commandover event socket
-					$tmp_cmd = 'api sofia profile '.$profile.' rescan';
-					$response = event_socket_request($fp, $tmp_cmd);
-					unset($tmp_cmd);
+					$cmd = 'api sofia profile '.$profile.' rescan';
+					$response = event_socket_request($fp, $cmd);
+					unset($cmd);
 				//close the connection
 					fclose($fp);
 			}
@@ -123,12 +114,8 @@ if (strlen($_GET["id"])>0) {
 }
 
 //redirect the users
-	require_once "resources/header.php";
-	echo "<meta http-equiv=\"refresh\" content=\"2;url=gateways.php\">\n";
-	echo "<div align='center'>\n";
-	echo $text['message-delete']."\n";
-	echo "</div>\n";
-	require_once "resources/footer.php";
+	$_SESSION["message"] = $text['message-delete'];
+	header("Location: gateways.php");
 	return;
 
 ?>
